add sql block to pipeline
