+++ mage_ai/frontend/components/FileBrowser/Folder.tsx
@@ -151,6 +151,7 @@ function Folder({
               );
             }
           }
+
           if (isPipelineFolder) {
             openPipeline(name);
           } else if (children) {
