+++ mage_ai/frontend/components/FileEditor/index.tsx
@@ -133,12 +133,13 @@ function FileEditor({
     setFilesTouched,
   ]);
 
-  const addToPipelineEl = fileExtension === FileExtensionEnum.PY
+  const addToPipelineEl = (fileExtension === FileExtensionEnum.PY || fileExtension === FileExtensionEnum.SQL)
     && getBlockType(file.path.split('/')) !== BlockTypeEnum.SCRATCHPAD && (
     <Spacing p={2}>
       <KeyboardShortcutButton
         inline
         onClick={() => addNewBlock({
+          language: FILE_EXTENSION_TO_LANGUAGE_MAPPING[fileExtension],
           name: getBlockUUID(file.path.split('/')),
           type: getBlockType(file.path.split('/')),
         })}
